#1-refactor(layout): migrate 5 layout components from JSX to TypeScript
