+++ it-viec-frontend/src/layout/LayoutEmployer/index.tsx
@@ -1,22 +1,23 @@
 import { Outlet } from "react-router-dom";
 import { Layout } from "antd";
 import "./LayoutEmployer.scss";
-import Header from "../../components/Header";
+import Header from "@/components/Header";
+import FooterComp from "@/components/Footer";
 
-import FooterComp from "../../components/Footer";
 const { Content } = Layout;
 
-function LayoutEmployer() {
+function LayoutEmployer(): JSX.Element {
   return (
     <>
       <Layout className="layout-default">
-        <Header type="employer"/>
+        <Header type="employer" />
         <Content className="content">
           <Outlet />
         </Content>
-        <FooterComp/>
+        <FooterComp />
       </Layout>
     </>
   );
 }
+
 export default LayoutEmployer;
